Update dual_echo
In numerical_model, there is a function named 'add_noise' which add noise using a normal distribution. The noise outside the phantom (or inside depending on the proton density) can be cancelled if we use the function 'threshold_masking', the output of this function is a mask for an input image by thresholding the signal. In 'multi_echo' the noise is cancelled using 'threshold_masking'. However in 'dual_echo' the noise is not cancelled because there is no use of 'threshold_masking' so we add it.